feat: Add event and currency type management
Enhance the Reservation Management feature by adding new services for
event and currency type management in the ReservationManagementController.
Update the FilterLookups and ReservationListFilter classes to include
new properties for currency types and events. Modify the Index view to
add dropdowns for selecting events and currency types, and improve the
filter layout. Update the _ReservationList view for better table
styling and include a new stylesheet for icons in the layout.